requirements.txt
matplotlib
numpy
tensorflow
pillow